backend/requirements.txt
Django==2.2.15
djangorestframework==3.11.1
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
